requirements.txt
asgiref==3.7.2
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
crispy-tailwind==0.5.0
cycler==0.12.1
Django==4.2.8
django-bootstrap5==23.3
django-crispy-forms==2.1
django-filter==23.5
django-social-share==2.3.0
et-xmlfile==1.1.0
folium==0.15.1
fonttools==4.48.1
geographiclib==2.0
geopy==2.4.1
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.0
xyzservices==2023.10.1
zipp==3.17.0
